Refactor: DRY up bunny DNS tests and clean up subdomainPreview plumbing
- Extract withMockBunnyCdnApi to #test-utils for shared use
- Replace 6 manual save/restore blocks in server tests with withMockBunnyCdnApi
- Hoist mockDnsZone helper to module scope for reuse across test suites
- Fold subdomainPreview into getAdvancedSettingsPageState, removing spread
- Extract isBunnyDnsEnabled local var to avoid duplicate calls
