+++ src/test-utils/index.ts
@@ -13,6 +13,7 @@ import { stub } from "@std/testing/mock";
 import forge from "node-forge";
 import { bracket } from "#fp";
 import type { SigningCredentials } from "#lib/apple-wallet.ts";
+import { bunnyCdnApi } from "#lib/bunny-cdn.ts";
 import { resetAllowedDomain } from "#lib/config.ts";
 import { getSessionCookieName, parseFlashValue } from "#lib/cookies.ts";
 import {
@@ -784,6 +785,25 @@ export const withMocks = async <
       }
     }
     await cleanup?.();
+  }
+};
+
+/** Temporarily replace bunnyCdnApi methods and restore after test */
+export const withMockBunnyCdnApi = async (
+  overrides: Partial<typeof bunnyCdnApi>,
+  fn: () => Promise<void>,
+): Promise<void> => {
+  const originals: Partial<typeof bunnyCdnApi> = {};
+  for (const key of Object.keys(overrides) as (keyof typeof bunnyCdnApi)[]) {
+    // deno-lint-ignore no-explicit-any
+    originals[key] = bunnyCdnApi[key] as any;
+    // deno-lint-ignore no-explicit-any
+    bunnyCdnApi[key] = overrides[key] as any;
+  }
+  try {
+    await fn();
+  } finally {
+    Object.assign(bunnyCdnApi, originals);
   }
 };
