+++ test/lib/server-settings-advanced.test.ts
@@ -23,6 +23,7 @@ import {
   setupEventAndLogin,
   testCookie,
   testCsrfToken,
+  withMockBunnyCdnApi,
   withMocks,
 } from "#test-utils";
 
@@ -702,146 +703,150 @@ describeWithEnv("server (admin settings-advanced)", { db: true }, () => {
 
         test("previews subdomain availability without save", async () => {
           setBunnyDnsEnv();
-          const original = bunnyCdnApi.checkSubdomainAvailable;
-          bunnyCdnApi.checkSubdomainAvailable = () =>
-            Promise.resolve({
-              ok: true as const,
-              available: true,
-              fullDomain: "myevent.tickets.example.com",
-            });
-          try {
-            const response = await handleRequest(
-              mockFormRequest(
-                "/admin/settings/host-subdomain",
-                {
-                  subdomain: "myevent",
-                  csrf_token: await testCsrfToken(),
-                },
-                await testCookie(),
-              ),
-            );
-            expect(response.status).toBe(302);
-            const location = response.headers.get("location")!;
-            expect(location).toContain("subdomain=myevent");
-            expect(location).toContain("form=settings-host-subdomain");
-          } finally {
-            bunnyCdnApi.checkSubdomainAvailable = original;
-          }
+          await withMockBunnyCdnApi(
+            {
+              checkSubdomainAvailable: () =>
+                Promise.resolve({
+                  ok: true as const,
+                  available: true,
+                  fullDomain: "myevent.tickets.example.com",
+                }),
+            },
+            async () => {
+              const response = await handleRequest(
+                mockFormRequest(
+                  "/admin/settings/host-subdomain",
+                  {
+                    subdomain: "myevent",
+                    csrf_token: await testCsrfToken(),
+                  },
+                  await testCookie(),
+                ),
+              );
+              expect(response.status).toBe(302);
+              const location = response.headers.get("location")!;
+              expect(location).toContain("subdomain=myevent");
+              expect(location).toContain("form=settings-host-subdomain");
+            },
+          );
         });
 
         test("preview returns error when availability check fails", async () => {
           setBunnyDnsEnv();
-          const original = bunnyCdnApi.checkSubdomainAvailable;
-          bunnyCdnApi.checkSubdomainAvailable = () =>
-            Promise.resolve({
-              ok: false as const,
-              error: "DNS zone error",
-            });
-          try {
-            const response = await handleRequest(
-              mockFormRequest(
-                "/admin/settings/host-subdomain",
-                {
-                  subdomain: "myevent",
-                  csrf_token: await testCsrfToken(),
-                },
-                await testCookie(),
-              ),
-            );
-            expect(response.status).toBe(502);
-          } finally {
-            bunnyCdnApi.checkSubdomainAvailable = original;
-          }
+          await withMockBunnyCdnApi(
+            {
+              checkSubdomainAvailable: () =>
+                Promise.resolve({
+                  ok: false as const,
+                  error: "DNS zone error",
+                }),
+            },
+            async () => {
+              const response = await handleRequest(
+                mockFormRequest(
+                  "/admin/settings/host-subdomain",
+                  {
+                    subdomain: "myevent",
+                    csrf_token: await testCsrfToken(),
+                  },
+                  await testCookie(),
+                ),
+              );
+              expect(response.status).toBe(502);
+            },
+          );
         });
 
         test("preview returns error when subdomain is taken", async () => {
           setBunnyDnsEnv();
-          const original = bunnyCdnApi.checkSubdomainAvailable;
-          bunnyCdnApi.checkSubdomainAvailable = () =>
-            Promise.resolve({
-              ok: true as const,
-              available: false,
-              fullDomain: "myevent.tickets.example.com",
-            });
-          try {
-            const response = await handleRequest(
-              mockFormRequest(
-                "/admin/settings/host-subdomain",
-                {
-                  subdomain: "myevent",
-                  csrf_token: await testCsrfToken(),
-                },
-                await testCookie(),
-              ),
-            );
-            expect(response.status).toBe(409);
-          } finally {
-            bunnyCdnApi.checkSubdomainAvailable = original;
-          }
+          await withMockBunnyCdnApi(
+            {
+              checkSubdomainAvailable: () =>
+                Promise.resolve({
+                  ok: true as const,
+                  available: false,
+                  fullDomain: "myevent.tickets.example.com",
+                }),
+            },
+            async () => {
+              const response = await handleRequest(
+                mockFormRequest(
+                  "/admin/settings/host-subdomain",
+                  {
+                    subdomain: "myevent",
+                    csrf_token: await testCsrfToken(),
+                  },
+                  await testCookie(),
+                ),
+              );
+              expect(response.status).toBe(409);
+            },
+          );
         });
 
         test("registers subdomain with save flag, saves to DB, and logs activity", async () => {
           setBunnyDnsEnv();
-          const original = bunnyCdnApi.registerBunnySubdomain;
-          bunnyCdnApi.registerBunnySubdomain = () =>
-            Promise.resolve({
-              ok: true as const,
-              fullDomain: "myevent.tickets.example.com",
-            });
-          try {
-            const response = await handleRequest(
-              mockFormRequest(
-                "/admin/settings/host-subdomain",
-                {
-                  subdomain: "myevent",
-                  save: "1",
-                  csrf_token: await testCsrfToken(),
-                },
-                await testCookie(),
-              ),
-            );
-            expectRedirectWithFlash(
-              "/admin/settings-advanced?form=settings-host-subdomain#settings-host-subdomain",
-              "Subdomain registered: myevent.tickets.example.com",
-            )(response);
-            expect(settings.bunnySubdomain).toBe("myevent.tickets.example.com");
-            const log = await getAllActivityLog();
-            expect(
-              log.some((e) =>
-                e.message.includes(
-                  "Host subdomain set to myevent.tickets.example.com",
+          await withMockBunnyCdnApi(
+            {
+              registerBunnySubdomain: () =>
+                Promise.resolve({
+                  ok: true as const,
+                  fullDomain: "myevent.tickets.example.com",
+                }),
+            },
+            async () => {
+              const response = await handleRequest(
+                mockFormRequest(
+                  "/admin/settings/host-subdomain",
+                  {
+                    subdomain: "myevent",
+                    save: "1",
+                    csrf_token: await testCsrfToken(),
+                  },
+                  await testCookie(),
                 ),
-              ),
-            ).toBe(true);
-          } finally {
-            bunnyCdnApi.registerBunnySubdomain = original;
-          }
+              );
+              expectRedirectWithFlash(
+                "/admin/settings-advanced?form=settings-host-subdomain#settings-host-subdomain",
+                "Subdomain registered: myevent.tickets.example.com",
+              )(response);
+              expect(settings.bunnySubdomain).toBe(
+                "myevent.tickets.example.com",
+              );
+              const log = await getAllActivityLog();
+              expect(
+                log.some((e) =>
+                  e.message.includes(
+                    "Host subdomain set to myevent.tickets.example.com",
+                  ),
+                ),
+              ).toBe(true);
+            },
+          );
         });
 
         test("returns error when registration fails", async () => {
           setBunnyDnsEnv();
-          const original = bunnyCdnApi.registerBunnySubdomain;
-          bunnyCdnApi.registerBunnySubdomain = () =>
-            Promise.resolve({
-              ok: false as const,
-              error: "DNS error",
-            });
-          try {
-            const response = await handleRequest(
-              mockFormRequest(
-                "/admin/settings/host-subdomain",
-                {
-                  subdomain: "myevent",
-                  save: "1",
-                  csrf_token: await testCsrfToken(),
-                },
-                await testCookie(),
-              ),
-            );
-            expect(response.status).toBe(502);
-          } finally {
-            bunnyCdnApi.registerBunnySubdomain = original;
-          }
+          await withMockBunnyCdnApi(
+            {
+              registerBunnySubdomain: () =>
+                Promise.resolve({ ok: false as const, error: "DNS error" }),
+            },
+            async () => {
+              const response = await handleRequest(
+                mockFormRequest(
+                  "/admin/settings/host-subdomain",
+                  {
+                    subdomain: "myevent",
+                    save: "1",
+                    csrf_token: await testCsrfToken(),
+                  },
+                  await testCookie(),
+                ),
+              );
+              expect(response.status).toBe(502);
+            },
+          );
         });
       });
     },
